perf(commitzen installation): install commitzen to assist in standardized workflow based on gitflow
Gitflow is a legacy Git workflow that in the beginning was an innovative and revolutionary strategy
for managing Git branches. Gitflow has lost popularity to trunk-based workflows, which are now
considered best practices for continuous modern software development and DevOps practices. Gitflow
can also be difficult to use with continuous integration/deployment. This post details Gitflow for
historical purposes.